cursor and timing things

+++ src/tty.h
@@ -10,4 +10,4 @@ void tty_write(size_t n, const char str[n]);
 void tty_printf(const char* format, ...);
 void tty_hangup(void);
 void tty_resize(int w, int h, Px pw, Px ph);
-bool tty_wait(Fd xfd, int timeout);
+bool tty_wait(Fd xfd, long long timeout);
